% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_psm}
\alias{get_psm}
\title{Posterior Similarity Matrix}
\usage{
get_psm(saveFileDir, trim = 0.5)
}
\arguments{
\item{saveFileDir}{directory containing result of a DPMUnc run}

\item{trim}{proportion of initial samples to discard (burn in)}
}
\value{
matrix with entries in \link{0,1} indicating the proportion of times each
pair of observations were in the same cluster
}
\description{
calculate a PSM from a DPMUnc run
}
\details{
uses a different formulation to calculate PSM to mcclust's comp.psm,
iterating over the distinct cluster labels instead of the observations,
which should be faster when the maximum number of clusters is not large
}
\author{
Chris Wallace
}
