% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_samples}
\alias{read_samples}
\title{Read samples}
\usage{
read_samples(saveFileDir)
}
\arguments{
\item{saveFileDir}{character vector of directories, either one directory per
chain run on the same data, or the "container" directory, holding one
subdirectory per chain which will be automatically found and read}
}
\value{
mcmc.list of samples, compatible with coda library
}
\description{
Once the samples are read, all functions in the coda library which accept an
mcmc.list can be used for the standard post-sampling diagnostics and
inference
}
\author{
Chris Wallace
}
