% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPMUnc.R
\name{scale_data}
\alias{scale_data}
\title{Scale observed data so variance of every variable is 1}
\usage{
scale_data(obsData, obsVars)
}
\arguments{
\item{obsData}{The observed data in matrix form (n observations x p variables)}

\item{obsVars}{The observed variances of the data (n observations x p variables)}
}
\value{
Named list containing scaled data and scaled variances
}
\description{
Scale observed data so variance of every variable is 1
}
\examples{
obsData <- matrix(rnorm(400, mean=0, sd=c(1,2,3,4)), ncol=4, byrow = TRUE)
obsVars <- matrix(rep(c(1, 4, 9, 16), 10), ncol=4, byrow = TRUE)
result = scale_data(obsData, obsVars)
# After scaling, var of every variable in data should be 1
apply(result$data, var, MARGIN=2)
# and the scaled variances will be approximately 1
apply(result$vars, mean, MARGIN=2)
}
