% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_mcmc}
\alias{read_mcmc}
\title{Read in MCMC output}
\usage{
read_mcmc(saveFileDir, burnin)
}
\arguments{
\item{saveFileDir}{a character vector of directories, each corresponding to
a single chain from the analysis of the same data}

\item{burnin}{what fraction of each chain to discard as burnin or number of obs to discard}
}
\value{
an mcmc.list
}
\description{
Reads in samples and stores as an mcmc.list compatible with the coda library.
}
\details{
This means output can be processed using using any coda function
The function ‘mcmc’ is used to create a Markov Chain Monte Carlo
object.  The input data are taken to be a vector, or a matrix with
one column per variable.
}
\author{
Chris Wallace
}
